requirements.txt
torch==1.8.1
gpytorch==1.4.2
numpy==1.19.5
pandas==1.1.5
scikit-learn==0.24.2
scipy==1.5.4
pillow==8.3.1

